Advent of Code - Day 19
Notes: for class Scanners function beaconsMatch, I added the combining of matching scanners distance data to the origin beacons after finding out that if a scanner matches with another scanner it will have twelve plus matches or no matches at all ( where if that wasn't the case, it would probably be best to do the matches in the mergeScanners function, which is where I originally was doing it until I found out about the zero or twelve plus matches. So I moved that step into beaconsMatch to save some processing time ).